Move require patch to preinstall.js
